Work on custom director support

+++ include/globals.hxx
@@ -11,7 +11,7 @@ namespace BetterSMS {
     extern bool sIsAudioStreaming;
     extern bool sIsAudioStreamAllowed;
     extern bool sIsDebugMode;
-    extern f32 sScreenWidth;
+    extern u32 sScreenWidth;
     extern f32 sFrameRate;
 
     bool isDebugMode();
